Cut the amount of checks in basic checker by 2
Just to be clear - I knew I could do this all along. The point was for this to be as least optimal as could be.
I'm making this change now to be able to compare this in higher numbers, since it's pretty much required for the other functions to properly compete.